// src/app/modules/users/user.type.ts
export type IUserFilterRequest = {
  searchTerm?: string | undefined
  firstName?: string | undefined
  middleName?: string | undefined
  lastName?: string | undefined
  uniqueId?: string | undefined
  role?: string | undefined
  nid?: string | undefined
  email?: string | undefined
  phoneNo?: string | undefined
  gender?: string | undefined
}